Added 'all packages - definition.py - illustrate' (di*) command.
Use (g di*) to illustrate all segments.

Candidacy protocol means that segments without changes will leave the
filesystem as-is.

NOTE: This means that you can run (di*) regularly as a type of scorewide test
while working.

Also good for ongoing scorewide tests are ...

    (s dt*)
    (s pt*)
    (g dc*)
    (g di*)

... and probably a few related commands.

These are all designed to allow for frequent runs as proof that nothing has
changed.

+++ scoremanager/idetools/SegmentPackageManager.py
@@ -184,7 +184,7 @@ class SegmentPackageManager(ScoreInternalPackageManager):
         self._open_versioned_file('illustration.ly')
 
     # TODO: refactor with MaterialPackageManager.illustrate_output_py()
-    def illustrate_definition_py(self):
+    def illustrate_definition_py(self, dry_run=False):
         r'''Illustrates ``definition.py``.
 
         Makes ``illustration.ly`` and ``illustration.pdf``.
@@ -224,6 +224,11 @@ class SegmentPackageManager(ScoreInternalPackageManager):
             self._path,
             'illustration.pdf',
             )
+        inputs, outputs = [], []
+        if dry_run:
+            inputs.append(self._definition_py_path)
+            outputs.append((illustration_ly_path, illustration_pdf_path))
+            return inputs, outputs
         with systemtools.FilesystemState(remove=temporary_files):
             shutil.copyfile(boilerplate_path, illustrate_path)
             with self._io_manager._silent():
